pyproject.toml
[project]
name = "datastar-tmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioquic>=1.2.0",
    "datastar-py>=0.6.0",
    "faker>=37.4.0",
    "jinja2>=3.1.6",
    "load-dotenv>=0.1.0",
    "psutil>=7.1.3",
    "requests>=2.32.5",
    "sanic[http3]>=25.3.0",
    "twitchio>=3.1.0",
]
